modify different page and structue folders
