--- src/components/Navbar.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-export default function Navbar() {
-  return (
-    <div>
-      <h1>i am Navbar </h1>
-      <p> Lorem ipsum dolor sit amet consectetur adipisicing elit. Rerum necessitatibus quia numquam blanditiis dolores quidem?</p>
-      <button>click me </button>
-    </div>
-  )
-}
